requirements.txt
tensorflow
keras
streamlit
matplotlib
opencv-python
numpy
pandas
pillow
